Move keystore to Docker volume
Generate keystore in to a Docker volume `keystore` and mount that at runtime. Switched to JRE image as previous commit removed the need to compile any Java code - just need access to keytool (and openssl). I'm guessing that the svirt_sandbox_file_t is not required when mounting a Docker volume but I have no environment in which to verify this. Likewise, in the absence of any end-to-end tests, my testing is limited to spinning this environment up on a native Linux environment and poking the UI a bit.

The move to Docker Compose V2 files opens the door to introducing additional overlay networks and removing links but that's another commit...

Signed-off-by: David Currie <david_currie@uk.ibm.com>