:wrench: Ensure startup is run, check versions
Very strange things can happen if you are starting with a
clean project, but really old images. It is also good to
be generally more helpful with regard to prereqs.

Changes made to docker scripts to check the version
at setup time. Change made to go-admin to ensure setup is
run.

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>