Add couch host and port for setup usage
development mode startup.sh scripts for couch today in player/map build an auth url to couchdb assuming the couch host will always be called couchdb and at the known port. This change adds that information to the main gameon.env so the scripts can stop making that assumption, allowing the host & port to be define as required per runtime environment.