Add in room service and modify yaml resources
This commit adds in instructions for deploying the gameon app in IBM Cloud Container service.
This commit also adds in the room service and room service environment variable in the configmap.
The generated certificate from openssl is now created as a secret instead of a configmap.
The yaml files using the certificate is now pointed to the new secret.
go-run.sh script should now create a secret from the generated certificate.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>